Focus the test on argsort.
